Add regression report classes for single-target and multi-target and aux class to measure time
